% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrumar_origem.R
\name{arrumar_origem}
\alias{arrumar_origem}
\title{arrumar_origem}
\usage{
arrumar_origem(paths)
}
\arguments{
\item{x}{lista de string.}
}
\description{
Recebe os caminhos (paths) dos arquivos no dir como parâmetro e
retorna uma string arrumada com a origem dos arquivos.
}
