% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidar_arquivos.R
\name{consolidar_arquivos}
\alias{consolidar_arquivos}
\title{consolidar_arquivos}
\usage{
consolidar_arquivos()
}
\arguments{
\item{x}{lista de string.}
}
\description{
Função sem passagem de parâmetros que retorna um arquivo .xlsx
}
