% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_arquivos.R
\name{ler_arquivos}
\alias{ler_arquivos}
\title{ler_arquivos}
\usage{
ler_arquivos(paths, n_sheet)
}
\arguments{
\item{x}{lista de string e uma constante.}
}
\description{
Recebe os caminhos (paths) dos arquivos no dir e o número de sheet como
parâmetro e retorna uma tibble arrumada.
}
